+++ src/GitHub_Updater/GitLab_API.php
@@ -18,9 +18,10 @@ if ( ! defined( 'WPINC' ) ) {
 }
 
 /**
+ * Class GitLab_API
+ *
  * Get remote data from a GitLab repo.
  *
- * Class    GitLab_API
  * @package Fragen\GitHub_Updater
  * @author  Andy Fragen
  */
@@ -28,7 +29,6 @@ class GitLab_API extends API {
 
 	/**
 	 * Holds loose class method name.
-	 *
 	 * @var null
 	 */
 	protected static $method = null;
@@ -107,6 +107,9 @@ class GitLab_API extends API {
 		$repo_type = $this->return_repo_type();
 		$response  = isset( $this->response['tags'] ) ? $this->response['tags'] : false;
 
+		if ( $this->exit_no_update( $response ) ) {
+			return false;
+		}
 
 		if ( ! $response ) {
 			$id           = $this->get_gitlab_id();
@@ -141,6 +144,20 @@ class GitLab_API extends API {
 	 */
 	public function get_remote_changes( $changes ) {
 		$response = isset( $this->response['changes'] ) ? $this->response['changes'] : false;
+
+		/*
+		 * Set response from local file if no update available.
+		 */
+		if ( ! $response && ! $this->can_update( $this->type )  ) {
+			$response = new \stdClass();
+			$content = $this->get_local_info( $this->type, $changes );
+			if ( $content ) {
+				$response->content = $content;
+				$this->set_transient( 'changes', $response );
+			} else {
+				$response = false;
+			}
+		}
 
 		if ( ! $response ) {
 			$id           = $this->get_gitlab_id();
@@ -182,6 +199,19 @@ class GitLab_API extends API {
 		}
 
 		$response = isset( $this->response['readme'] ) ? $this->response['readme'] : false;
+
+		/*
+		 * Set $response from local file if no update available.
+		 */
+		if ( ! $response && ! $this->can_update( $this->type )  ) {
+			$response = new \stdClass();
+			$content = $this->get_local_info( $this->type, 'readme.txt' );
+			if ( $content ) {
+				$response->content = $content;
+			} else {
+				$response = false;
+			}
+		}
 
 		if ( ! $response ) {
 			$id           = $this->get_gitlab_id();
@@ -239,7 +269,6 @@ class GitLab_API extends API {
 
 		$this->type->repo_meta = $response;
 		$this->_add_meta_repo_object();
-		$this->get_remote_branches();
 
 		return true;
 	}
@@ -252,6 +281,10 @@ class GitLab_API extends API {
 	public function get_remote_branches() {
 		$branches = array();
 		$response = isset( $this->response['branches'] ) ? $this->response['branches'] : false;
+
+		if ( $this->exit_no_update( $response, true ) ) {
+			return false;
+		}
 
 		if ( ! $response ) {
 			$id           = $this->get_gitlab_id();
@@ -420,6 +453,7 @@ class GitLab_API extends API {
 			if ( $this->type->repo === $project->name ) {
 				$id = $project->id;
 				$this->set_transient( 'projects', $response );
+				break;
 			}
 		}